refactor: download method
